Added support for LVGL Touch Screen
A touch screen can now be used in combination with LVGL and the pre-compiled tasmota32-lvgl.bin firmware. The heating controller will detect if LVGL is present and try to initialise the screen otherwise a basic i2c display will be used if connected. The UI has been designed for 320x240 resolution touch screen.